Possible behavior after ministep, simstep or twostep

ts_alternatives_ministep_beh() constructs the possible future behavior score after a ministep of ego. ts_alternatives_twostep_beh() constructs the possible future behavioral scores after two ministeps of two egos. ts_alternatives_simstep_beh() constructs all possible future behavioral scores (over the complete range of the behavioral variable) of one ego.Ego is thus allowed to jump from one extreme to the other

Usage

ts_alternatives_ministep_beh(beh, ego)

ts_alternatives_simstep_beh(beh, ego)

ts_alternatives_twostep_beh(beh, net, dist1 = NULL, modet1 = "degree")

Arguments

beh

numerical, vector representing the behavioral scores actors.

ego

numeric, value indicating ego (row number of net)

net

matrix, the adjacency matrix representing the relations between actors. Valid values are 0 and 1.

dist1

numeric, minimal path length between ego1 and ego2 at time1 in order to be allowed to start a coordination. If NULL all dyads are allowed to start a coordination (i.e. simultaneity).

modet1

string indicating the type of ties being evaluated at time1. "degree" considers all ties as undirected. "outdegree" only allows directed paths starting from ego1 and ending at ego2. "indegree" only allows directed paths starting from ego2 and ending at ego1. See: DETAILS.

Value

list, a list of the alternative vector representing te behavioral scores

See also

Examples

ccovar <- ts_prepdata(df_ccovar1)
ts_alternatives_ministep_beh(beh = ccovar[, "cov2"], ego = 3)
#> [[1]]
#>  [1]  0.4  0.4  0.4 -0.6  1.4 -1.6  0.4  3.4  0.4 -5.6
#> attr(,"mean")
#> [1] 1.6
#> attr(,"simMean")
#> [1] 0.7234568
#> attr(,"range")
#> [1] 9
#> attr(,"range2")
#> [1] -4  5
#> attr(,"simij")
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]        NA 1.0000000 0.8888889 0.8888889 0.8888889 0.7777778 1.0000000
#>  [2,] 1.0000000        NA 0.8888889 0.8888889 0.8888889 0.7777778 1.0000000
#>  [3,] 0.8888889 0.8888889        NA 0.7777778 1.0000000 0.6666667 0.8888889
#>  [4,] 0.8888889 0.8888889 0.7777778        NA 0.7777778 0.8888889 0.8888889
#>  [5,] 0.8888889 0.8888889 1.0000000 0.7777778        NA 0.6666667 0.8888889
#>  [6,] 0.7777778 0.7777778 0.6666667 0.8888889 0.6666667        NA 0.7777778
#>  [7,] 1.0000000 1.0000000 0.8888889 0.8888889 0.8888889 0.7777778        NA
#>  [8,] 0.6666667 0.6666667 0.7777778 0.5555556 0.7777778 0.4444444 0.6666667
#>  [9,] 1.0000000 1.0000000 0.8888889 0.8888889 0.8888889 0.7777778 1.0000000
#> [10,] 0.3333333 0.3333333 0.2222222 0.4444444 0.2222222 0.5555556 0.3333333
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6666667 1.0000000 0.3333333
#>  [2,] 0.6666667 1.0000000 0.3333333
#>  [3,] 0.7777778 0.8888889 0.2222222
#>  [4,] 0.5555556 0.8888889 0.4444444
#>  [5,] 0.7777778 0.8888889 0.2222222
#>  [6,] 0.4444444 0.7777778 0.5555556
#>  [7,] 0.6666667 1.0000000 0.3333333
#>  [8,]        NA 0.6666667 0.0000000
#>  [9,] 0.6666667        NA 0.3333333
#> [10,] 0.0000000 0.3333333        NA
#> 
#> [[2]]
#>  [1]  0.4  0.4  1.4 -0.6  1.4 -1.6  0.4  3.4  0.4 -5.6
#> attr(,"mean")
#> [1] 1.6
#> attr(,"simMean")
#> [1] 0.7234568
#> attr(,"range")
#> [1] 9
#> attr(,"range2")
#> [1] -4  5
#> attr(,"simij")
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]        NA 1.0000000 0.8888889 0.8888889 0.8888889 0.7777778 1.0000000
#>  [2,] 1.0000000        NA 0.8888889 0.8888889 0.8888889 0.7777778 1.0000000
#>  [3,] 0.8888889 0.8888889        NA 0.7777778 1.0000000 0.6666667 0.8888889
#>  [4,] 0.8888889 0.8888889 0.7777778        NA 0.7777778 0.8888889 0.8888889
#>  [5,] 0.8888889 0.8888889 1.0000000 0.7777778        NA 0.6666667 0.8888889
#>  [6,] 0.7777778 0.7777778 0.6666667 0.8888889 0.6666667        NA 0.7777778
#>  [7,] 1.0000000 1.0000000 0.8888889 0.8888889 0.8888889 0.7777778        NA
#>  [8,] 0.6666667 0.6666667 0.7777778 0.5555556 0.7777778 0.4444444 0.6666667
#>  [9,] 1.0000000 1.0000000 0.8888889 0.8888889 0.8888889 0.7777778 1.0000000
#> [10,] 0.3333333 0.3333333 0.2222222 0.4444444 0.2222222 0.5555556 0.3333333
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6666667 1.0000000 0.3333333
#>  [2,] 0.6666667 1.0000000 0.3333333
#>  [3,] 0.7777778 0.8888889 0.2222222
#>  [4,] 0.5555556 0.8888889 0.4444444
#>  [5,] 0.7777778 0.8888889 0.2222222
#>  [6,] 0.4444444 0.7777778 0.5555556
#>  [7,] 0.6666667 1.0000000 0.3333333
#>  [8,]        NA 0.6666667 0.0000000
#>  [9,] 0.6666667        NA 0.3333333
#> [10,] 0.0000000 0.3333333        NA
#> 
#> [[3]]
#>  [1]  0.4  0.4  2.4 -0.6  1.4 -1.6  0.4  3.4  0.4 -5.6
#> attr(,"mean")
#> [1] 1.6
#> attr(,"simMean")
#> [1] 0.7234568
#> attr(,"range")
#> [1] 9
#> attr(,"range2")
#> [1] -4  5
#> attr(,"simij")
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]        NA 1.0000000 0.8888889 0.8888889 0.8888889 0.7777778 1.0000000
#>  [2,] 1.0000000        NA 0.8888889 0.8888889 0.8888889 0.7777778 1.0000000
#>  [3,] 0.8888889 0.8888889        NA 0.7777778 1.0000000 0.6666667 0.8888889
#>  [4,] 0.8888889 0.8888889 0.7777778        NA 0.7777778 0.8888889 0.8888889
#>  [5,] 0.8888889 0.8888889 1.0000000 0.7777778        NA 0.6666667 0.8888889
#>  [6,] 0.7777778 0.7777778 0.6666667 0.8888889 0.6666667        NA 0.7777778
#>  [7,] 1.0000000 1.0000000 0.8888889 0.8888889 0.8888889 0.7777778        NA
#>  [8,] 0.6666667 0.6666667 0.7777778 0.5555556 0.7777778 0.4444444 0.6666667
#>  [9,] 1.0000000 1.0000000 0.8888889 0.8888889 0.8888889 0.7777778 1.0000000
#> [10,] 0.3333333 0.3333333 0.2222222 0.4444444 0.2222222 0.5555556 0.3333333
#>            [,8]      [,9]     [,10]
#>  [1,] 0.6666667 1.0000000 0.3333333
#>  [2,] 0.6666667 1.0000000 0.3333333
#>  [3,] 0.7777778 0.8888889 0.2222222
#>  [4,] 0.5555556 0.8888889 0.4444444
#>  [5,] 0.7777778 0.8888889 0.2222222
#>  [6,] 0.4444444 0.7777778 0.5555556
#>  [7,] 0.6666667 1.0000000 0.3333333
#>  [8,]        NA 0.6666667 0.0000000
#>  [9,] 0.6666667        NA 0.3333333
#> [10,] 0.0000000 0.3333333        NA
#>