Network Statistics

These functions calculate the respective network statistic for ego. When multiplied with the importance of each statistic (the 'parameters') this constitutes the network evaluation of ego. See: ts_eval().

Usage

ts_degree(net, ego)

ts_recip(net, ego)

ts_outAct(net, ego)

ts_inAct(net, ego)

ts_outPop(net, ego)

ts_inPop(net, ego)

ts_transTrip(net, ego)

ts_transMedTrip(net, ego)

ts_transRecTrip(net, ego)

ts_cycle3(net, ego)

ts_egoX(net, ego, cov)

ts_altX(net, ego, cov)

ts_diffX(net, ego, cov)

ts_simX(net, ego, cov)

ts_absDiffX(net, ego, cov)

ts_sameX(net, ego, cov)

ts_egoXaltX(net, ego, cov)

Arguments

net

matrix, the adjacency matrix representing the relations between actors. Valid values are 0 and 1.

ego

numeric, the ego for which we want to calculate the network statistic.

cov

numeric, covariate scores

Value

numeric value

Details

For examples on how to use these statistics see: vignette("1.Introduction_RsienaTwoStep", package="RsienaTwoStep").

For the mathematical definition of these network statistics see chapter 12 of the RSiena manual (Ripley et al. 2022) .

References

Ripley RM, Snijders TA, Boda Z, Vörös A, Preciado P (2022). Manual for SIENA version 4.0 (version August 11, 2022). http://www.stats.ox.ac.uk/~snijders/siena/RSiena_manual.pdf.

See also

Examples

ts_degree(net=ts_net1, ego=3)
#> [1] 2