Evaluation function

Evaluation function

Usage

ts_eval(net, ego, statistics, ccovar = NULL, parameters)

Arguments

net

matrix, the adjacency matrix representing the relations between actors. Valid values are 0 and 1.

ego

numeric, the ego for which we are going to calculate how its evaluates the network.

statistics

list of names of statistic functions the same length as parameters. See e.g. ts_degree() for a list of available statistic functions.

ccovar

data frame with named time-constant covariates

parameters

vector of numeric values the same length as statistics.

Details

For the mathematical definition of the evaluation function see chapter 12 of the RSiena manual ripley2022manualRsienaTwoStep.

References

ripley2022manualRsienaTwoStep

See also

Examples

ts_eval(net=ts_net1, ego=6, statistics=list(ts_degree, ts_recip),
parameters=c(-2,1))
#> [1] -4
ts_eval(net=ts_net1, ego=10, ccovar=df_ccovar1, statistics=list(ts_degree, ts_recip, ts_transTrip,
ts_transMedTrip, list(ts_egoX, "cov1")), parameters=c(-2,2,7,7,1))
#> [1] 11.4