Simulation of Network evolution

ts_sims is the workhorse function of the package RsienaTwoStep. It simulates the network evolution nsims times given an existing network net, the defined evaluation function (ts_eval()) by statistics and parameters and the average number of possible tie changes per actor as defined by rate.

Usage

ts_sims(
  nsims = 1000,
  parallel = FALSE,
  net,
  rate,
  statistics,
  parameters,
  p2step = c(0, 1, 0),
  chain = FALSE,
  dist1 = NULL,
  dist2 = NULL,
  modet1 = "degree",
  modet2 = "degree"
)

Arguments

nsims

numeric, number of simulations.

parallel

TRUE/FALSE

net

matrix, the adjacency matrix representing the relations between actors. Valid values are 0 and 1.

rate

numeric, the average number of possible tie-changes per actor in the simulation.

statistics,

list of names of statistic functions (see e.g. ts_degree() for a list of available functions)

parameters,

numeric vector the same length as parameters

p2step

numeric vector of length 3, setting the ratio of ministep, twostep and twoministeps.

chain

TRUE/FALSE, set to TRUE if you want to save all the subsequent networks (after the ministep or twostep) during the simulation. If FALSE only the end network is saved.

dist1

numeric, minimal path length between ego1 and ego2 at time1 in order to be allowed to start a coorporation. If NULL all dyads are allowed to start a cooperation.

dist2

numeric, minimal path length between ego1 and ego2 at time2 in order for twostep to be counted as coorporation.

modet1

string, indicating the type of ties being evaluated at time1. "degree" considers all ties as undirected. "outdegree" only allows directed paths starting from ego1 and ending at ego2. "indegree" only allows directed paths starting from ego2 and ending at ego2.

modet2

string, indicating the type of ties being evaluated at time2. "degree" considers all ties as undirected. "outdegree" only allows directed paths starting from ego1 and ending at ego2. "indegree" only allows directed paths starting from ego2 and ending at ego2.

Value

If chain=FALSE a list (of length nsims) of adjacency matrices representing the final network after the simulated evolution. If chain=TRUE a list of lists of adjacency matrices. Each inner list represents the complete network evolution of one simulation. The outer list refers to the simulation run (with length nsims).

Details

For examples on how to use ts_sims see: vignette("1.Introduction_RsienaTwoStep", package="RsienaTwoStep") or the package website. Before you set parallel to TRUE make sure to set-up a cluster with the package doParallel (see Examples).

See also

Examples

ts_sims(net=net2, nsims=2, rate=2, parallel=FALSE, statistics=list(ts_degree, ts_recip), parameters=c(-2,1), p2step=c(0,1,0))
#> [[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    1
#> [2,]    0    0    0    0    1
#> [3,]    0    0    0    0    1
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    1    0
#> 
#> [[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    1    0    0    0
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#>